Pulled getConnection and getConnectionObject out to a common file (#111)
* Pulled getConnection and getConnectionObject (now called getDBConnection) out to a common file
* Changed to use common as a local npm package
* Suppress eslint warning
