+++ src/api/lambdas/bedrock-api-backend/package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "aws-sdk": "^2.1275.0",
+    "bedrock_common": "file:../../../../bedrock_common",
     "mssql": "^6.2.1",
     "pg": "^8.3.0"
   },
