+++ src/bedrock_common/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "bedrock_common",
+  "version": "1.0.0",
+  "description": "Local package of common utility functions for Bedrock",
+  "main": "bedrock_common.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
